refactor: move splitByParentheses to separate file
